Add patch for liblog to add a function expected by proprietary blobs.
The proprietary blobs expect the "__xlog_buf_printf" function to be
provided by liblog. In Android 4.2 the liblog.so blob (along with
libcutils.so blob due to linkage issues) was used. While this was a
valid (although not very good) approach due to being from the same
version in Android 4.3 it is no longer possible (nor advisable) to keep
using a binary blob due to API changes. Therefore, liblog is now patched
to add the expected function.