Add support for NVRAM system.
The NVRAM system is a MediaTek specific system used by most of its
proprietary features to get its configuration parameters. The
configuration parameters are stored in the "nvram" partition (and maybe
in the "pro_info" partition; I am not sure of its purpose, but it is
used by the NVRAM system too), and the proprietary binary blobs retrieve
them through the NVRAM system, which takes care of reading the partition
and getting the needed data (or at least I think so; it is just based on
my observation of the behavior of the NVRAM system and the proprietary
binary blobs, but I can be wrong, as I do not have access to its source
code).

The "nvram_daemon" is run at boot, which creates the "/data/nvram"
directory and populates it with several configuration files (probably
just caching them for their later use). It does not contain all the
configuration files, though; proprietary binary blobs may ask for some
parameters that are not included in those files, in which case the
appropriate configuration files will be added to the directory. However,
the "nvram_daemon" should be run anyway at boot, as besides populating
some configuration files it also performs other initialization tasks,
like calibrating the accelerometer and gyroscope sensors based on the
configuration read.

In the official Fairphone ROM the nvram_daemon is executed under the
"nvram" user, which uid is "9997". This user is MediaTek specific and it
is not part of the standard Android system users. Although the NVRAM
system probably would work without problems under the "system" user (the
main difference would be just who is able to access "/data/nvram", as
"nvram_daemon" already has access to files in the "system" group) the
"android_filesystem_config.h" configuration file is patched to add the
"nvram" user anyway.

Also, the "nvram" user must be given permissions to set the "nvram_init"
system property, although in this case the "device_perms.h" trick
included in CyanogenMod can be used withouth having to patch the system.