Add HAL module and daemons for sensors.
The data for most sensors can be got using the kernel drivers and the
proprietary sensors.default.so HAL module alone. However, in the case of
the magnetometer, an extra daemon is needed. This daemon reads the
magnetometer data from the kernel, processes it, and then writes the new
data back to the kernel to be then read again from /dev/hwmsensor by the
HAL module along with the data from the other sensors.

Although the HAL module does not explicitly depend on the NVRAM system
it does implicitly, as nvram_daemon calibrates the accelerometer and
gyroscope sensor when started. If not done the sensors would work
anyway, but their values would be slightly off.