Add proprietary binary blobs for GPS.
The proprietary GPS support is provided by several components. There is
a main GPS daemon, mnld, which is started through an init service by the
GPS HAL module, gps.default.so. The GPS HAL module controls the daemon,
and the daemon provides the data to the module. Internally, the main GPS
daemon uses another daemon, libmnlp_mt6628, which is specific to the
actual wireless combo chip. This chip requires some firmware to work,
which is loaded when needed by yet a different daemon, 6620_launcher.

Only support for basic GPS is provided; other features like AGPS were
not included.