Add wrapper for proprietary MediaTek GPS HAL module.
The proprietary GPS HAL module from MediaTek was built against a
non-standard "hardware/gps.h" header. Therefore, when used in a system
with a standard header it does not work as expected. To solve this, a
GPS HAL module that wraps the MediaTek one and acts as an interface
between it and the standard system was added.