Add support to the Wi-Fi system for the MT6628 wlan kernel driver.
The Wi-Fi system expects that the Wi-Fi net interfaces are added and
removed when the Wi-Fi HAL loads and unloads the Wi-Fi kernel module.
However, this is not the case with the MT6628 wlan kernel driver.
Although the interfaces are removed when the driver is unloaded they are
not added automatically when the driver is loaded; to add the interfaces
some special commands have to be sent to the kernel once the driver was
loaded.

Due to this, the Wi-Fi HAL was patched to add support for adding and
removing the net interfaces. To do this, in a similar way to how the
firmware loader is handled, the Wi-Fi HAL now (optionally) starts an
init service (defined at build time) to add or remove the net interfaces
when it receives a request to load or unload the driver. The kernel
module, however, is loaded by init when the system starts and then it is
kept always loaded; only the service is used when needed to add or
remove the interfaces. The service, for its part, calls a helper
command, mt6628_wlan_iface_ctrl, which sends the appropriate commands to
the kernel to add and remove the interfaces.

As the patch touches files belonging to the build system, the build must
be stopped and then launched again when they are modified. The same
approach used in the device tree is used for the vendor tree.