Add support for wpa_supplicant.
The Android version of wpa_supplicant supports some extra commands
besides the default ones supported by the control interface. Those extra
commands are used, for example, to get the MAC address of the device.
However, the implementation for those commands must be provided by an
external static library specific to each device and it is not included
in the wpa_supplicant sources.

The implementation of that library for the Fairphone 1 was based on the
implementation for the qcwcn driver from Qualcomm provided in the AOSP
tree. Only the "wpa_driver_nl80211_driver_cmd" function is needed; other
functions, like "wpa_driver_set_p2p_ps", are needed only when
ANDROID_P2P is defined when building wpa_supplicant, but that happens
only when BOARD_WLAN_DEVICE is either "bcmdhd" or "qcwcn" instead of for
all Android devices.

To start the wpa_supplicant the Wi-Fi system delegates on the Wi-Fi HAL.
It will start the p2p_supplicant or the wpa_supplicant init service,
depending on whether the P2P feature is required or not. Note that the
p2p_supplicant service supports both Wi-Fi and Wi-Fi Direct; it is not
Wi-Fi Direct instead of plain Wi-Fi.

In any case, the Wi-Fi HAL takes care of creating the initial
configuration file for the wpa_supplicant interfaces based on the
template "/system/etc/wifi/wpa_supplicant.conf". The wpa_supplicant
sources provide a Makefile to generate the template as needed (that is,
taking into account the control interface to be used by the Wi-Fi HAL),
which must be included from a local Android.mk.

The control interface set for the main interface (wlan0) is the
wpa_wlan0 socket created by the init service; the other option would be
for wpa_supplicant to create the socket in the
"/data/system/wpa_supplicant" directory. The control interface for the
secondary interface (p2p0) is created in "/data/misc/wifi/sockets". The
Wi-Fi HAL along with the configuration template take care of most of the
details, although the directory must be created explicitly.

The wlan kernel driver does not support adding and removing virtual
interfaces using the "add_virtual_intf" operation, so the
"use_p2p_group_interface=1" option to create a new interface for each
group connection can not be used neither on the configuration files nor
in the parameters passed by the service to wpa_supplicant; only one
group connection at a time is supported (as it will have to use the main
p2p0 interface).