Add support for DHCP.
The libnetutils library starts the DHCP daemon when requested using the
"dhcpcd_XXX" and "iprenew_XXX" init services, where "XXX" is the
interface to use; in the case of P2P interfaces they all use the "p2p"
suffix, no matter if the interface itself is "p2p0" or a group
interface, like "p2p-p2p0-1" (although there are no group interfaces in
the case of the Fairphone 1).

Besides the configuration parameters included in the service itself the
"/system/etc/dhcpcd/dhcpcd.conf" configuration file is passed to the
DHCP daemon. No special configuration is needed for the Fairphone 1, so
instead of using a custom file in the Fairphone 1 tree the configuration
file is copied at build time from the default configuration file
provided by dhcpcd; this prevents having to update the local copy in the
Fairphone 1 tree when the dhcpcd default configuration file is updated.