Update Wi-Fi HAL patch to Android 4.3.
In CyanogenMod 10.2 (but not in AOSP 4.3 nor in later versions of
CyanogenMod), the execution flow inside wifi_load_driver was changed in
the commit e3167c611b, and the call to wifi_change_driver_state must be
done now inside the 'strcmp(FIRMWARE_LOADER,"") == 0' condition. The
reason is that the old path is never executed (nor built) due to
WIFI_FIRMWARE_LOADER being always defined at the start of wifi.c, even
if its defined to an empty value.