Disable nvram user patch for CyanogenMod 11.0.
CyanogenMod 11.0 includes the NVRAM user by default, so there is no need
to patch system/core. The patch is just disabled instead of removed,
however, because other Android 4.4 ROMs and later CyanogenMod versions
do not include the NVRAM user; it is provided only by CyanogenMod 11.0.

Also note that, although CyanogenMod 11.0 gives permissions to the
"nvram" user to set the "nvram_init" property it does so only when
MTK_HARDWARE is defined. That definition is used in several other areas
of the main tree to change the behavior of certain packages, so in order
to have a fine grained control the "device_perms.h" file is kept instead
of dropping it and defining MTK_HARDWARE.