Add support for uint32 GpsAidingData.
GpsAidingData was changed from uint16_t to uint32_t in CyanogenMod 11.0
to add room for extra flags. Moreover, the value for some flag was also
modified, so it is necessary to perform a type and value conversion from
the GpsAidingData used in the main GPS HAL to the GpsAidingData used in
the wrapped proprietary MediaTek GPS HAL module.

However, as this change is specific to CyanogenMod, the conversion is
optional and it has to be enabled through a build parameter. This makes
possible to still use the wrapper with the standard GPS HAL for Android
4.4, for example, in AOSP.