Add dummy implementation for new private functions.
In CyanogenMod 11.0 ANDROID_P2P is now defined when building
wpa_supplicant for all Android devices instead of just for certain ones,
so it is necessary to provide an implementation for all the private
functions defined when ANDROID_P2P is defined.

The new private functions require features not provided by the MT6628
wlan kernel driver, so their implementation is just a dummy
implementation to point out that they are not really implemented.
Fortunately, this does not seem to prevent the Wi-Fi and Wi-Fi Direct
features from working.